+++ src/components/Home/TopVoters.css
@@ -1,0 +1,22 @@
+.TopVoters thead .sub.header {
+  margin: 0;
+}
+
+.TopVoters thead .sub.header:first-child {
+  padding-left: 16px;
+}
+
+.TopVoters thead .sub.header:last-child {
+  padding-right: 15px;
+}
+
+.TopVoters .TopVotersRow__User {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  padding-left: 15px;
+}
+
+.TopVoters .TopVotersRow__User > .sub.header {
+  margin: 0 14px 0 0;
+}
